requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiohttp-retry==2.9.1
aiosignal==1.4.0
asgiref==3.11.0
attrs==25.4.0
autobahn==25.11.1
Automat==25.4.16
cbor2==5.7.1
certifi==2025.11.12
cffi==2.0.0
channels==4.3.2
channels_redis==4.3.0
charset-normalizer==3.4.4
colorama==0.4.6
constantly==23.10.4
cryptography==46.0.3
daphne==4.2.1
Django==6.0
django-cors-headers==4.9.0
django-formtools==2.5.1
django-otp==1.6.3
django-phonenumber-field==8.4.0
django-two-factor-auth==1.18.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
frozenlist==1.8.0
geographiclib==2.1
geopy==2.4.1
hyperlink==21.0.0
idna==3.11
Incremental==24.11.0
msgpack==1.1.2
multidict==6.7.0
packaging==25.0
propcache==0.4.1
psycopg2-binary==2.9.11
py-ubjson==0.16.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
PyJWT==2.10.1
pyOpenSSL==25.3.0
python-dotenv==1.2.1
qrcode==8.2
redis==7.1.0
requests==2.32.5
service-identity==24.2.0
sqlparse==0.5.4
stripe==14.0.1
twilio==9.8.8
Twisted==25.5.0
txaio==25.12.1
typing_extensions==4.15.0
tzdata==2025.2
ujson==5.11.0
urllib3==2.6.0
yarl==1.22.0
zope.interface==8.1.1
